put regimen class in sep file; continue to develop
