+++ inst/base/mrgsolve-evtools.h
@@ -66,81 +66,3 @@ void push(databox& self, mrg::evdata ev) {
 }
 
 
-namespace evt {
-
-class regimen {
-public:
-  regimen();
-  void init(databox& self);
-  void reset();
-  void dose();
-  void cmt(int cmt_);
-  int cmt() {return Cmt;}
-  void amt(double amt_);
-  double amt() {return Amt;}
-  void ii(double ii_);
-  double ii() {return Ii;}
-  void rate(double rate_);
-  double rate() {return Rate;}
-  void dc(double dose_stop_);
-  double dose_time;
-  double prev_dose_time;
-  double dose_stop;
-  
-private: 
-  databox* Self;
-  int Cmt;
-  double Ii;
-  double Amt; 
-  double Rate;
-};
-
-void regimen::reset() {
-  Cmt = 1;
-  Amt = 0.0;
-  Ii = 24.0;
-  Rate = 0.0;
-  dose_time = 0.0;
-  dose_stop = 1.0e9;
-  prev_dose_time = -1e9;
-}
-
-regimen::regimen() {
-  reset();
-}
-
-void regimen::init(databox& self_) {
-  reset();
-  Self = &self_;
-}
-
-void regimen::amt(double amt_) {
-  Amt = amt_;
-  return;
-}
-
-void regimen::ii(double ii_) {
-  Ii = ii_;
-  if(prev_dose_time > -1e9) {
-    dose_time = prev_dose_time + Ii;  
-  }
-  return;
-}
-
-void regimen::dose() {
-  if(!evt::near(Self->time, dose_time)) return;
-  if(Self->time > dose_stop) return;
-  mrg::evdata ev = evt::bolus(Amt, Cmt);
-  ev.rate = Rate;
-  Self->push(ev);
-  prev_dose_time = dose_time;
-  dose_time = dose_time + Ii;
-  return;
-}
-
-void regimen::dc(double dose_stop_) {
-  dose_stop = dose_stop_;
-  return;
-}
-
-}
